Refactor namespaces
