+++ src/net-questdb-client-tests/TcpTests.cs
@@ -30,8 +30,8 @@ using Org.BouncyCastle.Asn1.Sec;
 using Org.BouncyCastle.Crypto.Parameters;
 using Org.BouncyCastle.Math;
 using Org.BouncyCastle.Security;
-using QuestDB.Ingress;
-using QuestDB.Ingress.Utils;
+using QuestDB;
+using QuestDB.Utils;
 
 namespace net_questdb_client_tests;
