+++ src/net-questdb-client/Enums/ErrorCode.cs
@@ -23,14 +23,47 @@
  *
  ******************************************************************************/
 
-namespace QuestDB.Ingress;
+
+namespace QuestDB.Enums;
 
 /// <summary>
-///     Deprecated enum. 
+///     Standard error codes for QuestDB ILP clients.
 /// </summary>
-[Obsolete]
-public enum BufferOverflowHandling
+public enum ErrorCode
 {
-    SendImmediately,
-    Extend
+    /// The host, port, or interface was incorrect.
+    CouldNotResolveAddr,
+
+    /// Called methods in the wrong order. E.g. `symbol` after `column`.
+    InvalidApiCall,
+
+    /// A network error connecting or flushing data out.
+    SocketError,
+
+    /// The string or symbol field is not encoded in valid UTF-8.
+    /// 
+    /// *This error is reserved for the
+    /// [C and C++ API](https://github.com/questdb/c-questdb-client/).*
+    InvalidUtf8,
+
+    /// The table name or column name contains bad characters.
+    InvalidName,
+
+    /// The supplied timestamp is invalid.
+    InvalidTimestamp,
+
+    /// Error during the authentication process.
+    AuthError,
+
+    /// Error during TLS handshake.
+    TlsError,
+
+    /// The server does not support ILP-over-HTTP.
+    HttpNotSupported,
+
+    /// Error sent back from the server during flush.
+    ServerFlushError,
+
+    /// Bad configuration.
+    ConfigError
 }
