+++ src/net-questdb-client/Enums/TlsVerifyType.cs
@@ -24,46 +24,16 @@
  ******************************************************************************/
 
 
-namespace QuestDB.Ingress.Enums;
+// ReSharper disable InconsistentNaming
+
+namespace QuestDB.Enums;
 
 /// <summary>
-///     Standard error codes for QuestDB ILP clients.
+///     Enum for tls_verify.
+///     Defaults to 'on'.
 /// </summary>
-public enum ErrorCode
+public enum TlsVerifyType
 {
-    /// The host, port, or interface was incorrect.
-    CouldNotResolveAddr,
-
-    /// Called methods in the wrong order. E.g. `symbol` after `column`.
-    InvalidApiCall,
-
-    /// A network error connecting or flushing data out.
-    SocketError,
-
-    /// The string or symbol field is not encoded in valid UTF-8.
-    /// 
-    /// *This error is reserved for the
-    /// [C and C++ API](https://github.com/questdb/c-questdb-client/).*
-    InvalidUtf8,
-
-    /// The table name or column name contains bad characters.
-    InvalidName,
-
-    /// The supplied timestamp is invalid.
-    InvalidTimestamp,
-
-    /// Error during the authentication process.
-    AuthError,
-
-    /// Error during TLS handshake.
-    TlsError,
-
-    /// The server does not support ILP-over-HTTP.
-    HttpNotSupported,
-
-    /// Error sent back from the server during flush.
-    ServerFlushError,
-
-    /// Bad configuration.
-    ConfigError
+    unsafe_off,
+    on
 }
