+++ src/net-questdb-client/LineTcpSender.cs
@@ -32,9 +32,10 @@ using Org.BouncyCastle.Asn1.Sec;
 using Org.BouncyCastle.Crypto.Parameters;
 using Org.BouncyCastle.Math;
 using Org.BouncyCastle.Security;
+
 // ReSharper disable InconsistentNaming
 
-namespace QuestDB.Ingress;
+namespace QuestDB;
 
 [Obsolete("This has been superseded by ISender and the Sender factory.")]
 public class LineTcpSender : IDisposable
