+++ src/net-questdb-client/Senders/TcpSender.cs
@@ -30,13 +30,13 @@ using Org.BouncyCastle.Asn1.Sec;
 using Org.BouncyCastle.Crypto.Parameters;
 using Org.BouncyCastle.Math;
 using Org.BouncyCastle.Security;
-using QuestDB.Ingress.Enums;
-using QuestDB.Ingress.Utils;
-using Buffer = QuestDB.Ingress.Buffers.Buffer;
-using ProtocolType = QuestDB.Ingress.Enums.ProtocolType;
+using QuestDB.Enums;
+using QuestDB.Utils;
+using Buffer = QuestDB.Buffers.Buffer;
+using ProtocolType = QuestDB.Enums.ProtocolType;
 
 
-namespace QuestDB.Ingress.Senders;
+namespace QuestDB.Senders;
 
 /// <summary>
 ///     An implementation of <see cref="ISender"/> for TCP transport.
