+++ src/net-questdb-client/TlsMode.cs
@@ -23,20 +23,16 @@
  *
  ******************************************************************************/
 
-// ReSharper disable UnusedAutoPropertyAccessor.Global
-#pragma warning disable CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider declaring as nullable.
-namespace QuestDB.Ingress.Utils;
 
-// ReSharper disable once ClassNeverInstantiated.Global
-public record JsonErrorResponse
+namespace QuestDB;
+
+/// <summary>
+///     Deprecated TLS selection class.
+/// </summary>
+[Obsolete]
+public enum TlsMode
 {
-    public string code { get; init; }
-    public string message { get; init; }
-    public int line { get; init; }
-    public string errorId { get; init; }
-    
-    public override string ToString()
-    {
-        return $"\nServer Response (\n\tCode: `{code}`\n\tMessage: `{message}`\n\tLine: `{line}`\n\tErrorId: `{errorId}` \n)";
-    }
+    Enable,
+    AllowAnyServerCertificate,
+    Disable
 }
