+++ src/tcp-client-test/LineTcpSenderTests.cs
@@ -34,8 +34,7 @@ using Org.BouncyCastle.Asn1.Sec;
 using Org.BouncyCastle.Crypto.Parameters;
 using Org.BouncyCastle.Math;
 using Org.BouncyCastle.Security;
-using QuestDB.Ingress;
-using QuestDB.Ingress.Enums;
+using QuestDB;
 
 namespace tcp_client_test;
